# backend/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Rental Dashboard API",
    "description": "RESTful API for managing car rental operations including cars, bookings, customers, expenses, maintenance, and incidents.",
    "version": "1.0.0",
    "contact": {
      "name": "API Support"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3010",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token or API key authentication"
      }
    },
    "schemas": {
...
      "Booking": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "car_id": { "type": "integer" },
          "plate_number": { "type": "string" },
          "make": { "type": "string" },
          "model": { "type": "string" },
          "customer_name": { "type": "string" },
          "customer_phone": { "type": "string", "nullable": true },
          "customer_id_passport": { "type": "string", "nullable": true },
          "customer_id": { "type": "integer", "nullable": true },
          "start_date": { "type": "string", "format": "date" },
          "end_date": { "type": "string", "format": "date" },
          "status": { "type": "string", "enum": ["draft", "reserved", "confirmed", "active", "completed", "cancelled"] },
          "total_price": { "type": "number", "format": "float" },
          "deposit": { "type": "number", "format": "float" },
          "notes": { "type": "string", "nullable": true },
          "start_mileage": { "type": "integer", "nullable": true },
          "end_mileage": { "type": "integer", "nullable": true },
          "miles_driven": { "type": "integer" },
          "actual_return_date": { "type": "string", "format": "date", "nullable": true },
          "late_fee_amount": { "type": "number", "format": "float" },
          "created_at": { "type": "string", "format": "date-time" },
          "extras": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "extra_name": { "type": "string" },
                "extra_price": { "type": "number" }
              }
            }
          }
        }
      },
      "Customer": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "name": { "type": "string" },
          "email": { "type": "string", "format": "email", "nullable": true },
          "phone": { "type": "string", "nullable": true },
          "address": { "type": "string", "nullable": true },
          "id_number": { "type": "string", "nullable": true },
          "license_expiry": { "type": "string", "format": "date", "nullable": true },
          "created_at": { "type": "string", "format": "date-time" }
        }
      },
      "Expense": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "car_id": { "type": "integer" },
          "plate_number": { "type": "string" },
          "make": { "type": "string" },
          "model": { "type": "string" },
          "category": { "type": "string", "enum": ["maintenance", "insurance", "registration", "cleaning", "misc", "fuel"] },
          "amount": { "type": "number", "format": "float" },
          "description": { "type": "string", "nullable": true },
          "expense_date": { "type": "string", "format": "date" },
          "created_at": { "type": "string", "format": "date-time" }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/customers": {
      "get": {
        "summary": "List all customers",
        "tags": ["Customers"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "search", "in": "query", "schema": { "type": "string" }, "description": "Search in name, email, phone, or ID number" }
        ],
        "responses": {
          "200": {
            "description": "List of customers",
            "content": {
              "application/json": {
                "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Customer" } }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new customer",
        "tags": ["Customers"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name"],
                "properties": {
                  "name": { "type": "string", "minLength": 2, "maxLength": 100 },
                  "email": { "type": "string", "format": "email", "nullable": true },
                  "phone": { "type": "string", "minLength": 10, "maxLength": 20, "nullable": true },
                  "address": { "type": "string", "maxLength": 500, "nullable": true },
                  "id_number": { "type": "string", "maxLength": 50, "nullable": true },
                  "license_expiry": { "type": "string", "format": "date", "nullable": true }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Customer created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Customer" }
              }
            }
          },
          "400": { "description": "Invalid input" },
          "403": { "description": "Forbidden" }
        }
      }
    },
    "/api/customers/{id}": {
      "get": {
        "summary": "Get customer by ID with booking history",
        "tags": ["Customers"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": {
            "description": "Customer details with bookings",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "customer": { "$ref": "#/components/schemas/Customer" },
                    "bookings": { "type": "array", "items": { "$ref": "#/components/schemas/Booking" } }
                  }
                }
              }
            }
          },
          "404": { "description": "Customer not found" }
        }
      },
      "put": {
        "summary": "Update customer",
        "tags": ["Customers"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string" },
                  "email": { "type": "string", "format": "email", "nullable": true },
                  "phone": { "type": "string", "nullable": true },
                  "address": { "type": "string", "nullable": true },
                  "id_number": { "type": "string", "nullable": true },
                  "license_expiry": { "type": "string", "format": "date", "nullable": true }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Customer updated",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Customer" }
              }
            }
          },
          "400": { "description": "Invalid input" },
          "403": { "description": "Forbidden" },
          "404": { "description": "Customer not found" }
        }
      },
      "delete": {
        "summary": "Delete customer",
        "tags": ["Customers"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": {
            "description": "Customer deleted",
            "content": {
              "application/json": {
                "schema": { "type": "object", "properties": { "ok": { "type": "boolean" } } }
              }
            }
          },
          "403": { "description": "Forbidden - Admin only" },
          "404": { "description": "Customer not found" }
        }
      }
    },
...
    "/api/expenses": {
      "get": {
        "summary": "List all expenses",
        "tags": ["Expenses"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "car_id", "in": "query", "schema": { "type": "integer" } },
          { "name": "category", "in": "query", "schema": { "type": "string", "enum": ["maintenance", "insurance", "registration", "cleaning", "misc", "fuel"] } },
          { "name": "from", "in": "query", "schema": { "type": "string", "format": "date" } },
          { "name": "to", "in": "query", "schema": { "type": "string", "format": "date" } }
        ],
        "responses": {
          "200": {
            "description": "List of expenses",
            "content": {
              "application/json": {
                "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Expense" } }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new expense",
        "tags": ["Expenses"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["car_id", "category", "amount"],
                "properties": {
                  "car_id": { "type": "integer" },
                  "category": { "type": "string", "enum": ["maintenance", "insurance", "registration", "cleaning", "misc", "fuel"] },
                  "amount": { "type": "number", "minimum": 0 },
                  "description": { "type": "string", "nullable": true },
                  "expense_date": { "type": "string", "format": "date" }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Expense created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Expense" }
              }
            }
          },
          "400": { "description": "Invalid input" },
          "403": { "description": "Forbidden" }
        }
      }
    },
    "/api/expenses/{id}": {
      "get": {
        "summary": "Get expense by ID",
        "tags": ["Expenses"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": {
            "description": "Expense details",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Expense" }
              }
            }
          },
          "404": { "description": "Expense not found" }
        }
      },
      "put": {
        "summary": "Update expense",
        "tags": ["Expenses"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "car_id": { "type": "integer" },
                  "category": { "type": "string", "enum": ["maintenance", "insurance", "registration", "cleaning", "misc", "fuel"] },
                  "amount": { "type": "number", "minimum": 0 },
                  "description": { "type": "string", "nullable": true },
                  "expense_date": { "type": "string", "format": "date" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Expense updated",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Expense" }
              }
            }
          },
          "400": { "description": "Invalid input" },
          "403": { "description": "Forbidden" },
          "404": { "description": "Expense not found" }
        }
      },
      "delete": {
        "summary": "Delete expense",
        "tags": ["Expenses"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": {
            "description": "Expense deleted",
            "content": {
              "application/json": {
                "schema": { "type": "object", "properties": { "ok": { "type": "boolean" } } }
              }
            }
          },
          "403": { "description": "Forbidden - Admin only" },
          "404": { "description": "Expense not found" }
        }
      }
    },
...
  },
...
}